imap_email_ingestion_pipeline/requirements.txt
# Email Ingestion Pipeline Dependencies
# Core email processing
imaplib2==3.6
email-validator==2.0.0

# Document processing
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
tabula-py==2.9.0

# OCR engines
paddlepaddle==2.5.2
paddleocr==2.7.0.3
easyocr==1.7.0
pytesseract==0.3.10

# Image processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Database and state management
# Note: sqlite3 is built into Python, no installation needed

# Security and credentials
python-dotenv==1.0.0
cryptography==41.0.7

# File type detection
python-magic==0.4.27

# Monitoring and logging
prometheus-client==0.17.1

# HTML parsing and web scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Async HTTP client
aiohttp==3.8.6
aiofiles==23.2.0

# Enhanced PDF processing
pdfplumber==0.10.3

# Utilities
requests==2.31.0
tqdm==4.66.1

# Visualization libraries for dashboard
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
networkx>=2.6.0